All tests passing, updateData actually works, moved more shared code to base, shifted generate items back to each specific test
