+++ tests/test_dynamodb_full_table_sync.py
@@ -33,24 +33,29 @@ class DynamoDBFullTable(TestDynamoDBBase):
             'num_rows': 3531},
         ]
 
+    def generate_items(num_items):
+        serializer = TypeSerializer()
+        for i in range(num_items):
+            record = {
+                'int_id': i,
+                'decimal_field': decimal.Decimal(str(i) + '.00000000001'),
+                'string_field': random_string_generator(),
+                'byte_field': b'some_bytes',
+                'int_list_field': [i, i+1, i+2],
+                'int_set_field': set([i, i+1, i+2]),
+                'map_field': {
+                    'map_entry_1': 'map_value_1',
+                    'map_entry_2': 'map_value_2'
+                },
+                'string_list': [random_string_generator(), random_string_generator(), random_string_generator()],
+                'boolean_field': True,
+                'other_boolean_field': False,
+                'null_field': None
+            }
+            yield serializer.serialize(record)
+
     def name(self):
         return "tap_tester_dynamodb_full_table"
-
-    def tap_name(self):
-        return "tap-dynamodb"
-
-    def get_type(self):
-        return "platform.dynamodb"
-
-    def get_properties(self):
-        return {
-            "use_local_dynamo": 'true',
-            "account_id": "123123123123",
-            "region_name": "us-east-1"
-        }
-
-    def get_credentials(self):
-        return {}
 
     def test_run(self):
         conn_id = connections.ensure_connection(self)
